code/requirements.txt
numpy
pandas
torch
scikit-learn
python-binance
pandas-ta
TA-Lib
matplotlib